Remastering PoiController to accept image upload

+++ app/Controllers/Http/PoisController.ts
@@ -6,10 +6,12 @@ import TagPoi from 'App/Models/TagPoi';
 import PoiValidator from 'App/Validators/PoiValidator';
 import { PoiPreview, Response } from '../../../types/SharpObjects';
 import Database from '@ioc:Adonis/Lucid/Database';
-import Drive from '@ioc:Adonis/Core/Drive';
 import Env from '@ioc:Adonis/Core/Env';
 import fetch from 'cross-fetch';
 import PoiUpdateValidator from 'App/Validators/PoiUpdateValidator';
+import { PoiToStore } from '../../../types/SharpObjects';
+import Drive from '@ioc:Adonis/Core/Drive';
+import PoiValidatorForUpdate from 'App/Validators/PoiValidatorForUpdate';
 
 const MIN_PROBABILITY = 0.75;
 
@@ -56,7 +58,38 @@ export default class PoisController {
   public async store({ request, response }) {
     try {
       const data = await request.validate(PoiValidator);
-      const poi = await Poi.create(data);
+      const coverImage = request.file('image', {
+        extnames: ['jpg'],
+      });
+      if (!coverImage) {
+        return response.badRequest({
+          type: 'error',
+          content: 'no image with the key "image" sent',
+        });
+      }
+      if (!coverImage.isValid) {
+        return response.badRequest({
+          type: 'error',
+          content: coverImage.errors,
+        });
+      }
+
+      const today = Date.now();
+      coverImage.clientName = `${today}_${coverImage.clientName}`;
+      await coverImage.move(`${Env.get('URL_IMAGE')}`);
+      const url = `${Env.get('BASE_URL')}images/${coverImage.fileName}`;
+      const ObjectToStore: PoiToStore = {
+        area: data.area,
+        exhibition_number: data.exhibition_number,
+        title: data.title,
+        manufacturer: data.manufacturer,
+        periode: data.periode,
+        archived: data.archived,
+        image_url: url,
+        image_name: coverImage.clientName,
+        location: data.location,
+      };
+      const poi = await Poi.create(ObjectToStore);
       return poi;
     } catch (error) {
       return response.badRequest({
@@ -73,11 +106,30 @@ export default class PoisController {
    * @returns The updated POI
    */
   public async update({ params, request, response }) {
-    const data = await request.validate(PoiUpdateValidator);
+    const data = await request.validate(PoiValidatorForUpdate);
+    const coverImage = request.file('image', {
+      extnames: ['jpg'],
+    });
     const id = params.id;
     try {
       const poi = await Poi.findOrFail(id);
-      poi.merge(data);
+      await Drive.delete(`images/${poi.image_name}`);
+      const today = Date.now();
+      coverImage.clientName = `${today}_${coverImage.clientName}`;
+      await coverImage.move(`${Env.get('URL_IMAGE')}`);
+      const url = `${Env.get('BASE_URL')}images/${coverImage.fileName}`;
+      const ObjectToStore: PoiToStore = {
+        area: data.area,
+        exhibition_number: data.exhibition_number,
+        title: data.title,
+        manufacturer: data.manufacturer,
+        periode: data.periode,
+        archived: data.archived,
+        image_url: url,
+        image_name: coverImage.clientName,
+        location: data.location,
+      };
+      poi.merge(ObjectToStore);
       await poi.save();
       return poi;
     } catch (error) {
@@ -96,7 +148,18 @@ export default class PoisController {
   public async destroy({ params, response }) {
     try {
       const poi = await Poi.findOrFail(params.id);
-      return await poi.delete();
+      if (!(await Drive.exists(`images/${poi.image_name}`))) {
+        return response.badRequest({
+          type: 'error',
+          content: "Image doesn't exist",
+        });
+      }
+      await Drive.delete(`images/${poi.image_name}`);
+      await poi.delete();
+      return response.ok({
+        type: 'sucess',
+        content: 'resource deleted',
+      });
     } catch (error) {
       return response.badRequest({
         type: 'error',
